+++ packages/sdk/src/plt/index.ts
@@ -1,10 +1,9 @@
 export * from './types.js';
 
-export * as TokenEvent from './TokenEvent.js';
 export * as TokenId from './TokenId.js';
 export * as TokenModuleReference from './TokenModuleReference.js';
-export * as TokenModuleState from './TokenModuleState.js';
 export * as TokenAmount from './TokenAmount.js';
 export * as Token from './Token.js';
+export * as Cbor from './Cbor.js';
 
 export * as V1 from './v1/index.js';